refactor: group component index exported, revert example announcement
