+++ src/components/announcement/Announcements.tsx
@@ -1,24 +1,28 @@
-import { useEffect, useState } from 'react';
-import announcementApi from '@/api/announcementApi';
+import { AnnouncementBtn } from '@/components/buttons';
+import AnnouncementItem from '@/components/announcement/AnnouncementItem';
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
-import { AnnouncementBtn } from '@/components/buttons';
-import AnnouncementItem from '@/components/announcement/AnnouncementItem';
+import { useEffect, useState } from 'react';
+import announcementApi from '@/api/announcementApi';
 import { Announcement } from '@/types/Announcement';
 
-export default function Announcements() {
+// example은 어드민 페이지에서 신규 공지사항을 테스트하는 용도로 사용됩니다.
+export default function Announcements({ example }: { example?: Announcement }) {
   const [isOpen, setIsOpen] = useState(false);
   const [announcements, setAnnouncements] = useState<Announcement[]>([]);
   const currDate = new Date();
-
   useEffect(() => {
+    if (example) {
+      setAnnouncements([example]);
+      return;
+    }
     announcementApi.getAnnouncement({ page: 0, size: 30 }).then((res) => {
       setAnnouncements(res);
     });
-  }, []);
+  }, [example]);
 
   return (
     <DropdownMenu onOpenChange={(open) => setIsOpen(open)}>
