+++ src/components/group/group-utils/GroupCardContainer.tsx
@@ -2,7 +2,7 @@ import ProfileCard from '@/components/cards/ProfileCard';
 import { useCheckedStore, useCheckedUsersStore } from '@/lib/stores';
 import Group from '@/types/Group';
 
-export default function GroupCardContainer({ curGroup }: { curGroup: Group }) {
+export function GroupCardContainer({ curGroup }: { curGroup: Group }) {
   const { checked } = useCheckedStore();
   const { checkedUsers } = useCheckedUsersStore();
